Resolving first wave of comments by moving the route and altering the query
