+++ backend/conduit/profile/views.py
@@ -1,16 +1,19 @@
 # coding: utf-8
 
 from flask import Blueprint
-from flask_apispec import marshal_with
+from flask_apispec import marshal_with, use_kwargs
 from flask_jwt_extended import current_user, jwt_required, jwt_optional
 
 from conduit.exceptions import InvalidUsage
 from conduit.user.models import User
 from conduit.tags.models import Tags
+from conduit.articles.models import Article
+from marshmallow import fields
 
 from .serializers import profile_schema
 from conduit.tags.serializers import tags_schemas
 from conduit.user.serializers import followers_schema
+from conduit.articles.serializers import articles_schema
 
 blueprint = Blueprint('profiles', __name__)
 
@@ -76,3 +79,21 @@ def get_followings(username):
     if not user:
         raise InvalidUsage.user_not_found()
     return user.profile.follows
+
+
+#Route to return articles authored by current user
+@blueprint.route('/api/profile/articles', methods=('GET',))
+@jwt_required
+@use_kwargs({'type': fields.Str()})
+@marshal_with(articles_schema)
+def get_user_articles(type=None):
+    res = Article.query
+    if type != "all":
+        if type == "drafts":
+            res = res.filter_by(isPublished=False)
+        elif type == "published":
+            res = res.filter_by(isPublished=True)
+        else:
+            raise InvalidUsage.article_not_found()
+    res = res.join(Article.author).join(User).filter_by(username=current_user.username)
+    return res.all()
